requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
boto3==1.14.2
botocore==1.17.2
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.9.1
django-formtools==2.2
django-storages==1.9.1
django-subdomains==2.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-nested-routers==0.91
hiredis==1.0.1
humanize==2.5.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
Jinja2==2.11.2
jmespath==0.10.0
jwplatform==1.2.2
MarkupSafe==1.1.1
msgpack==0.6.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.0
stripe==2.48.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zope.interface==5.1.0
